Refactor NEAR integration and add provider modules
Refactored NEAR types from 'typeshare' to 'models', removed typeshare dependency, and added provider modules for balances, state, transactions, and preload. Updated NEAR RPC client to use generic Client and implemented new provider traits. Adjusted crate dependencies and feature flags, and updated usage in settings_chain and gemstone. Added tests and test data for NEAR transaction handling.

+++ crates/gem_near/src/models/account.rs
@@ -1,15 +1,11 @@
-use crate::typeshare::Int;
 use serde::{Deserialize, Serialize};
-use typeshare::typeshare;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
-#[typeshare(swift = "Sendable")]
 pub struct NearAccount {
     pub amount: String,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
-#[typeshare(swift = "Sendable")]
 pub struct NearAccountAccessKey {
-    pub nonce: Int,
+    pub nonce: i64,
 }
